Reimagining how the Docker entrypoint should work
By running an OpenSIPS container with Host Networking, we do not need
to query for the HOST_IP or the PUBLIC_IP.  This means we will have to
understand and configure OpenSIPS' global parameter "listen" before
starting the container.

The rsyslog is highly likely to be deprecated with a modern logging
system and using tail -f to keep this entrypoint running is not worthy
of greatness.

Instead, we check for an error-free OpenSIPS configuration file (that
is NOT the stock once from the repo) and that we are not about to ruin
someone's day by claiming an IP address that is already in use.